Contempt is eligible.
