+++ src/backend/tests/models/charge_types/test_civil_offense.py
@@ -41,66 +41,3 @@ def test_extradition():
         statute="00000336", level="Felony Unclassified", name="FUG/WA", disposition=Dispositions.DISMISSED
     )
     assert isinstance(charge.charge_type, CivilOffense)
-
-
-def test_contempt_of_court():
-    case = CaseSummaryFactory.create(type_status=["Contempt of Court", "Closed"])
-    charge = ChargeFactory.create(
-        case_number=case.case_number,
-        name="contempt of court",
-        statute="",
-        level="N/A",
-        disposition=Dispositions.DISMISSED,
-        violation_type=case.violation_type,
-    )
-
-    assert isinstance(charge.charge_type, CivilOffense)
-    assert charge.type_eligibility.status is EligibilityStatus.INELIGIBLE
-    assert charge.type_eligibility.reason == "Ineligible by omission from statute"
-
-
-def test_contempt_of_court_convicted():
-    case = CaseSummaryFactory.create(type_status=["Civil Offense", "Closed"])
-    charge = ChargeFactory.create(
-        case_number=case.case_number,
-        name="contempt of court",
-        statute="33065",
-        level="N/A",
-        disposition=Dispositions.CONVICTED,
-        violation_type=case.violation_type,
-    )
-
-    assert isinstance(charge.charge_type, CivilOffense)
-    assert charge.type_eligibility.status is EligibilityStatus.INELIGIBLE
-    assert charge.type_eligibility.reason == "Ineligible by omission from statute"
-
-
-def test_contempt_of_court_no_disposition():
-    case = CaseSummaryFactory.create(type_status=["Civil Offense", "Closed"])
-    charge = ChargeFactory.create(
-        case_number=case.case_number,
-        name="contempt of court",
-        statute="33015",
-        level="misdemeanor",
-        violation_type=case.violation_type,
-    )
-
-    assert isinstance(charge.charge_type, CivilOffense)
-    assert charge.type_eligibility.status is EligibilityStatus.INELIGIBLE
-    assert charge.type_eligibility.reason == "Ineligible by omission from statute"
-
-
-def test_contempt_of_court_unrecognized_disposition():
-    case = CaseSummaryFactory.create(type_status=["Civil Offense", "Closed"])
-    charge = ChargeFactory.create(
-        case_number=case.case_number,
-        name="contempt of court",
-        statute="33055",
-        level="violation",
-        disposition=Dispositions.UNRECOGNIZED_DISPOSITION,
-        violation_type=case.violation_type,
-    )
-
-    assert isinstance(charge.charge_type, CivilOffense)
-    assert charge.type_eligibility.status is EligibilityStatus.INELIGIBLE
-    assert charge.type_eligibility.reason == "Ineligible by omission from statute"
