+++ src/backend/tests/models/test_charge.py
@@ -11,6 +11,7 @@ from expungeservice.models.charge import ChargeType
 from expungeservice.models.disposition import DispositionCreator
 from tests.factories.charge_factory import ChargeFactory
 from tests.factories.case_factory import CaseSummaryFactory, CaseFactory
+from expungeservice.models.charge_types.contempt_of_court import ContemptOfCourt
 
 
 class Dispositions:
@@ -124,9 +125,9 @@ class TestChargeStatuteSectionAssignment(unittest.TestCase):
         assert section == "475B349"
 
 
-def test_nonblocking_charges_are_never_type_eligible():
+def test_nonblocking_charges_are_never_type_eligible_except_contempt_of_court():
     for charge_class in get_charge_classes():
-        if not charge_class().blocks_other_charges:
+        if not charge_class().blocks_other_charges and not isinstance(charge_class(), ContemptOfCourt):
             assert_charge_class_is_never_type_eligible(charge_class())
